pata_via kprobe: hack single channel registration
On an HP t5500 thin client the primary pata channel is used for the
internal flash and the second channel is not connected.  The onboard
ethernet controller uses IRQ 15, so this second IDE channel shares
the interrupt with the NIC.  Under network load irq/15-pata_via
actually shows up in top using non-trivial processor time.  This is
completely wasted time.  Use a kprobe to replace the id_table with
one that defines single channel use.  We can't use the subsystem
IDs to specifically identify this device for an upstream patch as
the sub-vendor/device matches vendor/device.  Perhaps DMI could be
used, but this works for now.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>